commit old changes

+++ source/core/IO.py
@@ -1,7 +1,8 @@
-# a file of local input functions
+# a file of local inout functions making the process easy and keeping the main file decluttered
 import os.path
-from . import syntax, log
+from source.core import syntax, log
 import yaml
+import json
 
 
 # simply prints the string output or creates a query in the terminal, also returns the inputted values
@@ -10,35 +11,28 @@ def say(output="", isQuestion=False, isLoop=False, syntaxChk=False, synType="", 
     if isQuestion:
 
         # seeing if the program wants to ask multiple queries
-        # if isLoop, then the output value should be an array and not a string
         if isLoop:
             loopVector = len(output)
             count = 0
-            usrSay = []  # this is the returned list from the keyboard
+            usrSay = []  # this is the returned list
             while loopVector > count:  # while the # of questions is greater than the responses, ask a question
-                usrSay.append(input(output[count])) # store the user's response to the printed question to the end of usrSay
-
-                # send the data to the log file for this instance
+                usrSay.append(input(output[count]))
                 log.log(output[count], "output")
                 log.log(usrSay[count], "input")
                 count += 1
-            # if there is syntax to be checked for, the syntax library will be called and the applicable syntax will be initialized and returned
+
             if syntaxChk:
                 return syntax.adv(usrSay, synType, isLoop)
-            # otherwise return the raw output
             else:
                 return usrSay
-        # if the output is only to be printed once
+
         else:
-            # logging
             log.log(output, "output")
-            
-            # again checking for syntax and calling the syntax library if it is True
+
             if syntaxChk:
                 response = syntax.adv(input(output), synType, isLoop)  # says the output and returns the prompt
                 log.log(response, "input")
                 return response
-            # otherwise, return the raw output
             else:
                 response = input(output)
                 log.log(response, "input")
@@ -70,7 +64,6 @@ def say(output="", isQuestion=False, isLoop=False, syntaxChk=False, synType="", 
 
 # WRITE TO A YAML FILE
 def yamlWrite(value, element, File, isLoop=False):
-    # check if there is data prestored in File
     data = yamlRead(File, element, True)
     if data is None:
         data = {}
@@ -109,6 +102,55 @@ def yamlRead(File, element, update=False, elements=1):
             if update:
                 count = 0
                 while elements > count:
+                    data.update({element[count]: data[element[count]]})
+                    count += 1
+
+                return data
+            else:
+                return data[element]
+
+# write into an element(s) of a json file
+def jsonWrite(value, element, File, isLoop=False):
+    # blah blah blah
+    data = jsonRead(File, element, True)
+    if data is None:
+        data = {}
+        if isLoop and len(element) == len(value):
+            count = 0
+            while len(element) > count:
+                data.update({element[count]: value[count]})
+                count += 1
+
+        else:
+            data.update({element: value})
+    else:
+        if isLoop:
+            count = 0
+            while len(element) > count:
+                data.update({element[count]: value[count]})
+                count += 1
+
+        else:
+            data.update({element: value})
+
+    with open(File, "w") as file:
+        log.log(element, "wfile", File)
+        json.dump(data, file)
+
+
+# read from a json file
+def jsonRead(File, element, update=False, elements=1):
+    # more blah
+    with open(File, "r") as file:
+        data = yaml.safe_load(file)
+        log.log(element, "rfile", File)
+        if data is None:
+            return None
+
+        else:
+            if update:
+                count = 0
+                while elements > count:
                     data.update({element: data[element]})
                     count += 1
 
@@ -122,6 +164,7 @@ def mkFile(File):
         log.log(File, "mkfile")
         return os.path.isfile(File)
 
+
 def mkDir(dir):
     os.mkdir(dir)
     log.log(dir, "mkdir")
@@ -129,35 +172,52 @@ def mkDir(dir):
 
 
 def fileWrite(value, File, isLoop=False, overwrite=False, update=False, isLog=False):
-    if overwrite and not update:
-        if isLoop:
-            count = 0
-            with open(File, "w") as file:
-                while len(value) > count:
-                    file.write(value[count] + "\n")
-                    if not isLog:
-                        log.log(value[count], "wfile", File)
-                    count += 1
-        else:
-            with open(File, "w") as file:
-                file.write(value)
-    elif not overwrite and not update:
-        if isLoop:
-            count = 0
-            with open(File, "a") as file:
-                while len(value) > count:
-                    file.write(value[count] + "\n")
-                    if not isLog:
-                        log.log(value[count], "wfile", File)
-                    count += 1
-        else:
-            with open(File, "a") as file:
-                file.write(value)
-                if not isLog:
-                    log.log(value, "wfile", File)
+    
+    if fileExists(File):
 
-    if update:
-        print()
+        if overwrite and not update:
+            if isLoop:
+                count = 0
+                with open(File, "w") as file:
+                    while len(value) > count:
+                        file.write(value[count] + "\n")
+                        if not isLog:
+                            log.log(value[count], "wfile", File)
+                        count += 1
+            else:
+                with open(File, "w") as file:
+                    file.write(value)
+        elif not overwrite and not update:
+            if isLoop:
+                count = 0
+                with open(File, "a") as file:
+                    while len(value) > count:
+                        file.write(value[count] + "\n")
+                        if not isLog:
+                            log.log(value[count], "wfile", File)
+                        count += 1
+            else:
+                with open(File, "a") as file:
+                    file.write(value)
+                    if not isLog:
+                        log.log(value, "wfile", File)
+
+        if update:
+            print()
+
+
+# reads a file and outputs an array
+def fileRead(File):
+    if fileExists(File):
+        with open(File, "r") as file:
+            out = []
+            for line in file:
+                out.append(line)
+
+        log.log(File, "rfile")
+        return out
+    else:
+        return []
 
 def rmFile(File):
     try:
@@ -167,12 +227,40 @@ def rmFile(File):
         return False
 
 
-# reads a file and outputs an array
-def fileRead(File):
-    with open(File, "r") as file:
-        out = []
-        for line in file:
-            out.append(line)
+def fileExists(File, logging=True):
+        
+    exists = os.path.isfile(File)
+    if logging:
+        if not exists:
+            say("File does not exist.")
+            log.log(f"Nonexistent file: {File}", "err")
 
-    log.log(File, "rfile")
-    return out
+    return exists
+
+def checksum(localFile, remoteFile):
+
+    match = True
+
+    if fileExists(localFile, False):
+
+        dataArray = fileRead(localFile)
+        localData = ""
+
+        for line in dataArray:
+            localData += line
+
+    else:
+
+        localData = ""
+
+    localData.encode()
+
+    remoteData = requests.get(remoteFile)
+
+    localmd5 = hashlib.md5(localData).hexdigest()
+    remotemd5 = hashlib.md5(remoteData.content).hexdigest()
+
+    if localmd5 != remotemd5:
+        match = False
+
+    return match
